mods/sector: move sector geometry out of the game core
Resolves the `// TODO: mods/sector` markers from #293. The RoomIntrinsics
sector fields move to a schema registrant in mods/sector, the geometry
helpers in game/room/sector.ts move whole, and GameMap's '#sectors' becomes
the mod-owned iterateSectors(world). The built World format is unchanged.
