+++ .changeset/sector-mod-extraction.md
@@ -1,0 +1,5 @@
+---
+"xxscreeps": patch
+---
+
+Move sector geometry out of the game core into a sector mod.
