+++ packages/xxscreeps/mods/modern/deposit/index.ts
@@ -6,6 +6,7 @@ export const manifest: Manifest = {
 		'xxscreeps/mods/modern/factory',
 		'xxscreeps/mods/classic/harvestable',
 		'xxscreeps/mods/classic/resource',
+		'xxscreeps/mods/sector',
 	],
 	provides: [ 'backend', 'constants', 'game', 'main', 'processor', 'schema', 'test' ],
 	types,
