+++ packages/xxscreeps/mods/modern/deposit/test.ts
@@ -8,8 +8,8 @@ import * as C from 'xxscreeps/game/constants/index.js';
 import { Game } from 'xxscreeps/game/index.js';
 import { createRoomObject } from 'xxscreeps/game/object.js';
 import { RoomPosition } from 'xxscreeps/game/position.js';
-import { makeSectorRadiusPredicate } from 'xxscreeps/game/room/sector.js';
 import { create as createCreep } from 'xxscreeps/mods/classic/creep/creep.js';
+import { makeSectorRadiusPredicate } from 'xxscreeps/mods/sector/sector.js';
 import { deterministicRandomForTesting } from 'xxscreeps/test/fixtures.js';
 import { testWorld } from 'xxscreeps/test/import.js';
 import { assert, describe, simulate, test } from 'xxscreeps/test/index.js';
