+++ packages/xxscreeps/mods/sector/index.ts
@@ -1,0 +1,5 @@
+import type { Manifest } from 'xxscreeps/config/mods.js';
+
+export const manifest: Manifest = {
+	provides: [ 'schema', 'test' ],
+};
